+++ source/extensions/omni.flux.feature_flags.window/omni/flux/feature_flags/window/tests/__init__.py
@@ -15,7 +15,4 @@
 * limitations under the License.
 """
 
-__all__ = ["FeatureFlagsCore", "FeatureFlag"]
-
-from .core import FeatureFlagsCore
-from .items import FeatureFlag
+from .e2e.test_window import TestFeatureFlagsWindow