+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - REMIX-3583: Added a Feature Flags system
 - REMIX-3567: Enable Sentry for built versions
 - REMIX-3113: Parallel process count dropdown for ingestion
+- REMIX-3583: Added tests for the Feature Flags system
 
 ### Changed
 - Updated runtime to 0.6.0-rc2