+++ source/extensions/omni.flux.feature_flags.window/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.1]
+### Added
+- Added E2E tests for the window
+
 ## [1.0.0]
 ### Added
 - Init commit.